refactor: centralize Settings state with React Context
Refs #193
