+++ src/renderer/src/components/settings/GamesSection.tsx
@@ -1,22 +1,14 @@
 import { GAMES } from '../../lib/config'
+import { useSettings } from './SettingsContext'
 
 interface GamesSectionProps {
   open: boolean
-  gamePaths: Record<string, string>
-  gameIcons: Record<string, string>
   onOpenChange: (open: boolean) => void
-  onBrowse: (key: string) => void
-  onGamePathChange: (key: string, path: string) => void
 }
 
-export function GamesSection({
-  open,
-  gamePaths,
-  gameIcons,
-  onOpenChange,
-  onBrowse,
-  onGamePathChange
-}: GamesSectionProps) {
+export function GamesSection({ open, onOpenChange }: GamesSectionProps) {
+  const { gamePaths, gameIcons, onBrowse, onGamePathChange } = useSettings()
+
   return (
     <section className="space-y-4">
       <button
@@ -88,7 +80,7 @@ export function GamesSection({
                   />
 
                   <button
-                    onClick={() => onBrowse(game.key)}
+                    onClick={() => onBrowse(game.key, true)}
                     className="accent-surface-action action-hover-scale cursor-pointer shrink-0 rounded-xl px-4 py-2 text-xs font-semibold"
                   >
                     Browse
